Extract DelimiterStatement to reduce duplicate logic for MySQL and doris
